cloud/metadata: extract imds client
